+++ examples/resources/tls_cert_request/resource.tf
@@ -1,5 +1,4 @@
 resource "tls_cert_request" "example" {
-  key_algorithm   = "ECDSA"
   private_key_pem = file("private_key.pem")
 
   subject {