+++ internal/provider/resource_self_signed_cert_test.go
@@ -19,7 +19,7 @@ func TestSelfSignedCert(t *testing.T) {
 		Providers: testProviders,
 		Steps: []r.TestStep{
 			{
-				Config: selfSignedCertConfig(1, 0),
+				Config: selfSignedCertConfig(1, 0, false),
 				Check: func(s *terraform.State) error {
 					gotUntyped := s.RootModule().Outputs["key_pem_1"].Value
 					got, ok := gotUntyped.(string)
@@ -129,7 +129,6 @@ func TestSelfSignedCert(t *testing.T) {
                         subject {
                             serial_number = "42"
                         }
-                        key_algorithm = "RSA"
                         validity_period_hours = 1
                         allowed_uses = []
                         private_key_pem = <<EOT
@@ -209,6 +208,120 @@ EOT
 	})
 }
 
+// TODO Remove this as part of https://github.com/hashicorp/terraform-provider-tls/issues/174
+func TestSelfSignedCert_HandleKeyAlgorithmDeprecation(t *testing.T) {
+	r.UnitTest(t, r.TestCase{
+		Providers: testProviders,
+		Steps: []r.TestStep{
+			{
+				Config: selfSignedCertConfig(1, 0, true),
+				Check: func(s *terraform.State) error {
+					gotUntyped := s.RootModule().Outputs["key_pem_1"].Value
+					got, ok := gotUntyped.(string)
+					if !ok {
+						return fmt.Errorf("output for \"key_pem_1\" is not a string")
+					}
+
+					if !strings.HasPrefix(got, "-----BEGIN CERTIFICATE----") {
+						return fmt.Errorf("key is missing cert PEM preamble")
+					}
+					block, _ := pem.Decode([]byte(got))
+					cert, err := x509.ParseCertificate(block.Bytes)
+					if err != nil {
+						return fmt.Errorf("error parsing cert: %s", err)
+					}
+					if expected, got := "2", cert.Subject.SerialNumber; got != expected {
+						return fmt.Errorf("incorrect subject serial number: expected %v, got %v", expected, got)
+					}
+					if expected, got := "example.com", cert.Subject.CommonName; got != expected {
+						return fmt.Errorf("incorrect subject common name: expected %v, got %v", expected, got)
+					}
+					if expected, got := "Example, Inc", cert.Subject.Organization[0]; got != expected {
+						return fmt.Errorf("incorrect subject organization: expected %v, got %v", expected, got)
+					}
+					if expected, got := "Department of Terraform Testing", cert.Subject.OrganizationalUnit[0]; got != expected {
+						return fmt.Errorf("incorrect subject organizational unit: expected %v, got %v", expected, got)
+					}
+					if expected, got := "5879 Cotton Link", cert.Subject.StreetAddress[0]; got != expected {
+						return fmt.Errorf("incorrect subject street address: expected %v, got %v", expected, got)
+					}
+					if expected, got := "Pirate Harbor", cert.Subject.Locality[0]; got != expected {
+						return fmt.Errorf("incorrect subject locality: expected %v, got %v", expected, got)
+					}
+					if expected, got := "CA", cert.Subject.Province[0]; got != expected {
+						return fmt.Errorf("incorrect subject province: expected %v, got %v", expected, got)
+					}
+					if expected, got := "US", cert.Subject.Country[0]; got != expected {
+						return fmt.Errorf("incorrect subject country: expected %v, got %v", expected, got)
+					}
+					if expected, got := "95559-1227", cert.Subject.PostalCode[0]; got != expected {
+						return fmt.Errorf("incorrect subject postal code: expected %v, got %v", expected, got)
+					}
+
+					if expected, got := 2, len(cert.DNSNames); got != expected {
+						return fmt.Errorf("incorrect number of DNS names: expected %v, got %v", expected, got)
+					}
+					if expected, got := "example.com", cert.DNSNames[0]; got != expected {
+						return fmt.Errorf("incorrect DNS name 0: expected %v, got %v", expected, got)
+					}
+					if expected, got := "example.net", cert.DNSNames[1]; got != expected {
+						return fmt.Errorf("incorrect DNS name 0: expected %v, got %v", expected, got)
+					}
+
+					if expected, got := 2, len(cert.IPAddresses); got != expected {
+						return fmt.Errorf("incorrect number of IP addresses: expected %v, got %v", expected, got)
+					}
+					if expected, got := "127.0.0.1", cert.IPAddresses[0].String(); got != expected {
+						return fmt.Errorf("incorrect IP address 0: expected %v, got %v", expected, got)
+					}
+					if expected, got := "127.0.0.2", cert.IPAddresses[1].String(); got != expected {
+						return fmt.Errorf("incorrect IP address 0: expected %v, got %v", expected, got)
+					}
+
+					if expected, got := 2, len(cert.URIs); got != expected {
+						return fmt.Errorf("incorrect number of URIs: expected %v, got %v", expected, got)
+					}
+					if expected, got := "spiffe://example-trust-domain/ca", cert.URIs[0].String(); got != expected {
+						return fmt.Errorf("incorrect URI 0: expected %v, got %v", expected, got)
+					}
+					if expected, got := "spiffe://example-trust-domain/ca2", cert.URIs[1].String(); got != expected {
+						return fmt.Errorf("incorrect URI 1: expected %v, got %v", expected, got)
+					}
+
+					if expected, got := 2, len(cert.ExtKeyUsage); got != expected {
+						return fmt.Errorf("incorrect number of ExtKeyUsage: expected %v, got %v", expected, got)
+					}
+					if expected, got := x509.ExtKeyUsageServerAuth, cert.ExtKeyUsage[0]; got != expected {
+						return fmt.Errorf("incorrect ExtKeyUsage[0]: expected %v, got %v", expected, got)
+					}
+					if expected, got := x509.ExtKeyUsageClientAuth, cert.ExtKeyUsage[1]; got != expected {
+						return fmt.Errorf("incorrect ExtKeyUsage[1]: expected %v, got %v", expected, got)
+					}
+
+					if expected, got := x509.KeyUsageKeyEncipherment|x509.KeyUsageDigitalSignature, cert.KeyUsage; got != expected {
+						return fmt.Errorf("incorrect KeyUsage: expected %v, got %v", expected, got)
+					}
+
+					// This time checking is a bit sloppy to avoid inconsistent test results
+					// depending on the power of the machine running the tests.
+					now := time.Now()
+					if cert.NotBefore.After(now) {
+						return fmt.Errorf("certificate validity begins in the future")
+					}
+					if now.Sub(cert.NotBefore) > (2 * time.Minute) {
+						return fmt.Errorf("certificate validity begins more than two minutes in the past")
+					}
+					if cert.NotAfter.Sub(cert.NotBefore) != time.Hour {
+						return fmt.Errorf("certificate validity is not one hour")
+					}
+
+					return nil
+				},
+			},
+		},
+	})
+}
+
 func TestAccSelfSignedCertRecreatesAfterExpired(t *testing.T) {
 	oldNow := overridableTimeFunc
 	var previousCert string
@@ -217,7 +330,7 @@ func TestAccSelfSignedCertRecreatesAfterExpired(t *testing.T) {
 		PreCheck:  setTimeForTest("2019-06-14T12:00:00Z"),
 		Steps: []r.TestStep{
 			{
-				Config: selfSignedCertConfig(10, 2),
+				Config: selfSignedCertConfig(10, 2, false),
 				Check: func(s *terraform.State) error {
 					gotUntyped := s.RootModule().Outputs["key_pem_1"].Value
 					got, ok := gotUntyped.(string)
@@ -229,7 +342,7 @@ func TestAccSelfSignedCertRecreatesAfterExpired(t *testing.T) {
 				},
 			},
 			{
-				Config: selfSignedCertConfig(10, 2),
+				Config: selfSignedCertConfig(10, 2, false),
 				Check: func(s *terraform.State) error {
 					gotUntyped := s.RootModule().Outputs["key_pem_1"].Value
 					got, ok := gotUntyped.(string)
@@ -247,7 +360,7 @@ func TestAccSelfSignedCertRecreatesAfterExpired(t *testing.T) {
 			},
 			{
 				PreConfig: setTimeForTest("2019-06-14T19:00:00Z"),
-				Config:    selfSignedCertConfig(10, 2),
+				Config:    selfSignedCertConfig(10, 2, false),
 				Check: func(s *terraform.State) error {
 					gotUntyped := s.RootModule().Outputs["key_pem_1"].Value
 					got, ok := gotUntyped.(string)
@@ -265,7 +378,7 @@ func TestAccSelfSignedCertRecreatesAfterExpired(t *testing.T) {
 			},
 			{
 				PreConfig: setTimeForTest("2019-06-14T21:00:00Z"),
-				Config:    selfSignedCertConfig(10, 2),
+				Config:    selfSignedCertConfig(10, 2, false),
 				Check: func(s *terraform.State) error {
 					gotUntyped := s.RootModule().Outputs["key_pem_1"].Value
 					got, ok := gotUntyped.(string)
@@ -294,7 +407,7 @@ func TestAccSelfSignedCertNotRecreatedForEarlyRenewalUpdateInFuture(t *testing.T
 		PreCheck:  setTimeForTest("2019-06-14T12:00:00Z"),
 		Steps: []r.TestStep{
 			{
-				Config: selfSignedCertConfig(10, 2),
+				Config: selfSignedCertConfig(10, 2, false),
 				Check: func(s *terraform.State) error {
 					gotUntyped := s.RootModule().Outputs["key_pem_1"].Value
 					got, ok := gotUntyped.(string)
@@ -306,25 +419,7 @@ func TestAccSelfSignedCertNotRecreatedForEarlyRenewalUpdateInFuture(t *testing.T
 				},
 			},
 			{
-				Config: selfSignedCertConfig(10, 3),
-				Check: func(s *terraform.State) error {
-					gotUntyped := s.RootModule().Outputs["key_pem_1"].Value
-					got, ok := gotUntyped.(string)
-					if !ok {
-						return fmt.Errorf("output for \"key_pem_1\" is not a string")
-					}
-
-					if got != previousCert {
-						return fmt.Errorf("certificate updated even though still time until early renewal")
-					}
-
-					previousCert = got
-					return nil
-				},
-			},
-			{
-				PreConfig: setTimeForTest("2019-06-14T16:00:00Z"),
-				Config:    selfSignedCertConfig(10, 3),
+				Config: selfSignedCertConfig(10, 3, false),
 				Check: func(s *terraform.State) error {
 					gotUntyped := s.RootModule().Outputs["key_pem_1"].Value
 					got, ok := gotUntyped.(string)
@@ -342,7 +437,25 @@ func TestAccSelfSignedCertNotRecreatedForEarlyRenewalUpdateInFuture(t *testing.T
 			},
 			{
 				PreConfig: setTimeForTest("2019-06-14T16:00:00Z"),
-				Config:    selfSignedCertConfig(10, 9),
+				Config:    selfSignedCertConfig(10, 3, false),
+				Check: func(s *terraform.State) error {
+					gotUntyped := s.RootModule().Outputs["key_pem_1"].Value
+					got, ok := gotUntyped.(string)
+					if !ok {
+						return fmt.Errorf("output for \"key_pem_1\" is not a string")
+					}
+
+					if got != previousCert {
+						return fmt.Errorf("certificate updated even though still time until early renewal")
+					}
+
+					previousCert = got
+					return nil
+				},
+			},
+			{
+				PreConfig: setTimeForTest("2019-06-14T16:00:00Z"),
+				Config:    selfSignedCertConfig(10, 9, false),
 				Check: func(s *terraform.State) error {
 					gotUntyped := s.RootModule().Outputs["key_pem_1"].Value
 					got, ok := gotUntyped.(string)
@@ -374,7 +487,6 @@ func TestAccSelfSignedCertSetSubjectKeyID(t *testing.T) {
 					subject {
 						serial_number = "42"
 					}
-					key_algorithm = "RSA"
 					validity_period_hours = 1
 					allowed_uses = []
 					set_subject_key_id = true
@@ -415,7 +527,6 @@ func TestAccSelfSignedCert_InvalidConfigs(t *testing.T) {
 						subject {
 							common_name = "common test cert"
 						}
-						key_algorithm = "RSA"
 						validity_period_hours = 1
 						allowed_uses = [
 							"not_valid"
@@ -435,7 +546,6 @@ func TestAccSelfSignedCert_InvalidConfigs(t *testing.T) {
 						subject {
 							common_name = "common test cert"
 						}
-						key_algorithm = "RSA"
 						validity_period_hours = -1
 						allowed_uses = [
 						]
@@ -454,7 +564,6 @@ func TestAccSelfSignedCert_InvalidConfigs(t *testing.T) {
 						subject {
 							common_name = "common test cert"
 						}
-						key_algorithm = "RSA"
 						validity_period_hours = 20
 						early_renewal_hours = -10
 						allowed_uses = [
@@ -472,7 +581,12 @@ func TestAccSelfSignedCert_InvalidConfigs(t *testing.T) {
 	})
 }
 
-func selfSignedCertConfig(validity uint32, earlyRenewal uint32) string {
+func selfSignedCertConfig(validity, earlyRenewal uint32, setKeyAlgorithm bool) string {
+	keyAlgorithmCfg := ""
+	if setKeyAlgorithm {
+		keyAlgorithmCfg = `key_algorithm = "RSA"`
+	}
+
 	return fmt.Sprintf(`
                     resource "tls_self_signed_cert" "test1" {
                         subject {
@@ -512,7 +626,7 @@ func selfSignedCertConfig(validity uint32, earlyRenewal uint32) string {
                             "client_auth",
                         ]
 
-                        key_algorithm = "RSA"
+                        %s
                         private_key_pem = <<EOT
 %s
 EOT
@@ -520,5 +634,5 @@ EOT
                     output "key_pem_1" {
                         value = tls_self_signed_cert.test1.cert_pem
                     }
-                `, validity, earlyRenewal, testPrivateKeyPEM)
+                `, validity, earlyRenewal, keyAlgorithmCfg, testPrivateKeyPEM)
 }